third_party/WebKit/LayoutTests/accessibility/textarea-caret-position-expected.txt

This tests that caret position is reported correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textareaAccessible.selectionStart is 0
PASS textareaAccessible.selectionEnd is 0
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 1
PASS textareaAccessible.selectionEnd is 1
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 2
PASS textareaAccessible.selectionEnd is 2
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 3
PASS textareaAccessible.selectionEnd is 3
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 4
PASS textareaAccessible.selectionEnd is 4
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 5
PASS textareaAccessible.selectionEnd is 5
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 6
PASS textareaAccessible.selectionEnd is 6
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 0
PASS textareaAccessible.selectionStart is 7
PASS textareaAccessible.selectionEnd is 7
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 8
PASS textareaAccessible.selectionEnd is 8
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 9
PASS textareaAccessible.selectionEnd is 9
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 10
PASS textareaAccessible.selectionEnd is 10
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 11
PASS textareaAccessible.selectionEnd is 11
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 12
PASS textareaAccessible.selectionEnd is 12
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 13
PASS textareaAccessible.selectionEnd is 13
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 14
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 15
PASS textareaAccessible.selectionEnd is 15
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 16
PASS textareaAccessible.selectionEnd is 16
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 17
PASS textareaAccessible.selectionEnd is 17
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 18
PASS textareaAccessible.selectionEnd is 18
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 19
PASS textareaAccessible.selectionEnd is 19
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 20
PASS textareaAccessible.selectionEnd is 20
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 20
PASS textareaAccessible.selectionEnd is 20
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE

 
